Move BotProfile to separate module
