+++ chess_trainer/trainer.py
@@ -17,8 +17,8 @@ try:  # optional chess engine library
     import chess.engine
 except Exception:  # pragma: no cover - optional dependency
     chess = None
-from dataclasses import dataclass, field
-from typing import List
+
+from .bot_profile import BotProfile
 
 load_dotenv()  # read token from environment if available
 API_TOKEN = os.getenv("LICHESS_BOT_TOKEN")
@@ -26,7 +26,6 @@ API_TOKEN = os.getenv("LICHESS_BOT_TOKEN")
 """Main bot training loop and event handling."""
 
 from . import openings_explorer
-
 STOCKFISH_PATH = "/usr/games/stockfish"
 # TODO check if the binary actually exists in this path, if not, do "which stockfish" and use that path instead
 
@@ -40,131 +39,6 @@ if berserk is not None and API_TOKEN:
 else:  # pragma: no cover - allows running tests without optional deps
     session = client = None
 
-white_openings = ["e4 - King's Pawn Game",
-                  "d4 - Queen's Pawn Game",
-                  "c4 - English Opening",
-                  "d4 d5 c4 - Queen's Gambit",
-                  "d4 d5 Bf4 - Queen's Pawn Game: Accelerated London System",
-                  "e4 e5 Nc3 - Vienna Game",
-                  "e4 e5 f4 - King's Gambit"
-                  ]
-black_openings = ["e4 e5 - King's Pawn Game",
-                  "e4 c5 - Sicilian Defense",
-                  "e4 d5 - Scandinavian Defense",
-                  "e4 e6 - French Defense",
-                  "e4 c6 - Caro-Kann Defense",
-                  "d4 Nf6 - Indian Defense",
-                  "d4 g6 - Queen's Pawn Game: Modern Defense"
-                  ]
-
-
-@dataclass
-class BotProfile:
-    chosen_white: List[str] = field(default_factory=list)
-    chosen_black: List[str] = field(default_factory=list)
-    our_color: bool = True # True = White, False = Black
-    opp_rating: int = 1500
-    challenge_rating: int = 1500
-    challenge: int = 100
-
-    def get_openings_choice_from_user(self):
-        def choose(options, color_name):
-            while True:
-                print(f"Select your {color_name} openings by number (comma-separated):")
-                print("  0. No preference")
-                for idx, opening in enumerate(options, start=1):
-                    print(f"  {idx}. {opening}")
-                user_input = input(f"Your {color_name} choices: ").strip()
-
-                # default or explicit “no pref”
-                # pick the "top 3" (according to me)
-                # so we don't play random garbage openings with early queen out, etc.
-                if user_input == "" or user_input == "0":
-                    return options[:3]
-
-                # parse comma-separated ints
-                picks = []
-                try:
-                    picks = [int(tok) for tok in user_input.split(",") if tok.strip()]
-                except ValueError:
-                    print(" -> Please enter only numbers, separated by commas.")
-                    continue
-
-                # allow 0 but reject negatives or > len(options)
-                if any(p < 0 or p > len(options) for p in picks):
-                    print(f" -> Each number must be between 0 and {len(options)}.")
-                    continue
-
-                # any zero means no preference, so pick the "top 3" (according to me)
-                # so we don't play random garbage openings with early queen out, etc.
-                if 0 in picks:
-                    return options[:3]
-
-                # dedupe & map to openings
-                seen = set()
-                selection = []
-                for p in picks:
-                    if p not in seen:
-                        seen.add(p)
-                        selection.append(options[p - 1])
-                return selection
-
-        self.chosen_white = choose(white_openings, "White")
-        self.chosen_black = choose(black_openings, "Black")
-        # self.challenge = int(input(f"Relative to your ELO, what should the BOT's ELO be? ").strip())
-        # new challenge-rating prompt with default = 0
-        while True:
-            user_input = input("Relative to your ELO, what should the BOT's ELO be? (+/- 100): ").strip()
-            if user_input in ("", "0"):
-                self.challenge = 0
-                break
-            try:
-                val = int(user_input)
-                self.challenge = val
-                break
-            except ValueError:
-                print(" -> Please enter a valid integer.")
-                continue
-
-    def determine_color_and_opp_rating(self, start):
-        """
-        {
-            'id': '4FOwDIIn',
-            'variant': {'key': 'standard', 'name': 'Standard', 'short': 'Std'},
-            'speed': 'rapid',
-            'perf': {'name': 'Rapid'},
-            'rated': False,
-            'createdAt': datetime.datetime(2025, 6, 19, 22, 25, 48, 138000, tzinfo=datetime.timezone.utc),
-            'white': {'id': 'wolfpacktwentythree', 'name': 'WolfpackTwentyThree', 'title': None, 'rating': 1778},
-            'black': {'id': 'chess-trainer-bot', 'name': 'chess-trainer-bot', 'title': 'BOT', 'rating': 2000, 'provisional': True},
-            'initialFen': 'startpos',
-            'clock': {'initial': 600000, 'increment': 5000},
-            'type': 'gameFull',
-            'state': {'type': 'gameState', 'moves': '', 'wtime': 600000, 'btime': 600000, 'winc': 5000, 'binc': 5000, 'status': 'started'}
-        }
-        """
-        white_player = start["white"]
-        black_player = start["black"]
-
-        # Determine our color and opponent rating
-        if white_player["id"] == OUR_NAME:
-            self.our_color = chess.WHITE
-            self.opp_rating = black_player["rating"]
-        else:
-            self.our_color = chess.BLACK
-            self.opp_rating = white_player["rating"]
-        self.challenge_rating = self.opp_rating + self.challenge
-        # return our_color, opp_rating
-
-    @staticmethod
-    def strip_opening_name(opening: str) -> str:
-        if "-" not in opening:
-            return opening
-        return opening.split("-", 1)[1].strip()
-
-    def get_clean_openings(self):
-        return ([self.strip_opening_name(o) for o in self.chosen_white],
-                [self.strip_opening_name(o) for o in self.chosen_black])
 
 ############################################### Core Bot Logic ####################################
